refactor(visual-recognition): remove compatibility layer for visual recognition
BREAKING CHANGE: Deprecated methods in Visual Recognition are no longer available. Changed parameter names are no longer interally corrected.

To migrate your code, use the methods and parameters currently available with the service as documented here: https://www.ibm.com/watson/developercloud/visual-recognition/api/v3/node.html?node
